:recycle: refactor: FE refactoring
