+++ src/components/Chatting/components/Messages/ChattingMessages.module.css
@@ -1,15 +1,31 @@
 .messagesContainer {
     flex: 1;
     overflow-y: auto;
-    width: 60%;
+    width: 100%;
+    padding-right: 20%;
+    padding-left: 20%;
     min-width: 600px;
     display: flex;
     margin: 0 auto;
-    padding: 0 20px;
     flex-direction: column;
     overflow-x: hidden;
     overflow-y: auto;
     box-sizing: border-box;
+}
+
+.messagesContainer::-webkit-scrollbar {
+    width: 10px;
+}
+.messagesContainer::-webkit-scrollbar-thumb {
+    background-color: #543fd6;
+    border-radius: 10px;
+    background-clip: padding-box;
+    border: 2px solid transparent;
+}
+.messagesContainer::-webkit-scrollbar-track {
+    background-color: rgba(84, 63, 214, 0.5);
+    border-radius: 10px;
+    box-shadow: inset 0px 0px 5px white;
 }
 
 .message {
