+++ src/components/SharingPrompt/Prompt/modal/DetailModal/PromptDetailModal.module.css
@@ -19,7 +19,7 @@
     padding: 2vw;
 
     width: 50vw;
-    height: 80vh;
+    height: auto;
     border-radius: 20px;
     background: #fff;
     box-shadow: 0px 0px 20px 0px rgba(0, 0, 0, 0.1);
@@ -27,7 +27,7 @@
 }
 
 .textSection {
-    height: 30vh;
+    gap: 2vh;
     display: flex;
     flex-direction: column;
     justify-content: space-between;
@@ -39,7 +39,6 @@
 
 .explainSection {
     width: 100%;
-    height: 10vh;
     overflow: auto;
     word-break: break-all;
 }
@@ -61,7 +60,6 @@
 
 .previewSectionText {
     width: 100%;
-    height: 10vh;
     overflow: auto;
     word-break: break-all;
 }
@@ -84,7 +82,6 @@
 .blockSection {
     display: flex;
     flex-direction: column;
-    margin-left: auto;
-    margin-right: auto;
-    height: 30vh;
+    margin: 3vh auto;
+    height: auto;
 }
