+++ src/components/SharingPrompt/Prompt/modal/DetailModal/components/ButtonSection.jsx
@@ -2,13 +2,13 @@ import React from "react";
 import styles from "./ButtonSection.module.css";
 import DeletePromptModal from "../../DeleteModal/DeletePromptModal";
 import {useRecoilState, useRecoilValue} from "recoil";
-import {myPageState} from "../../../../../recoil/community/myPageRecoilState";
+import {myPageState} from "../../../../../../recoil/community/myPageRecoilState";
 import PostFixOrShare from "../../FixAndShareModal/PostFixOrShare";
-import {useModalStack} from "../../../../../hooks/useModalStack";
-import {useCommunityHooks} from "../../../../../api/community/community";
-import {useMyPageHooks} from "../../../../../api/community/myPage";
-import {makePromptDetailState, stateChange} from "../../../../../recoil/community/communityRecoilState";
-import ModalButton from "../../../../common/ModalButton";
+import {useModalStack} from "../../../../../../hooks/useModalStack";
+import {useCommunityHooks} from "../../../../../../api/community/community";
+import {useMyPageHooks} from "../../../../../../api/community/myPage";
+import {makePromptDetailState, stateChange} from "../../../../../../recoil/community/communityRecoilState";
+import ModalButton from "../../../../../common/ModalButton";
 import {useNavigate} from "react-router-dom";
 import {t} from "i18next";
