+++ src/components/SharingPrompt/Prompt/modal/DetailModal/components/ButtonSection.module.css
