+++ src/components/SharingPrompt/Prompt/modal/DetailModal/components/TopSection.module.css
@@ -9,6 +9,7 @@
 .methodSection {
     grid-column: 1;
     justify-self: start;
+    text-align: center;
 }
 
 .likeSection {
