+++ src/pages/communityPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import SharePromptList from "../components/SharingPrompt/Prompt/SharePromptList";
-import FilterSection from "../components/SharingPrompt/community/FilterSection";
-import PageBar from "../components/PageBar/PageBar";
+import FilterSection from "../components/SharingPrompt/FilterSection/FilterSection";
+import PageBar from "../components/SharingPrompt/PageBar/PageBar";
 
 function CommunityPage() {
     return(
